Refactor endpoint to get users for an auth service
It's now possible to get users for every authentication backend instead
of only the active one.

Refs Graylog2/graylog-plugin-enterprise#1912
