+++ graylog2-server/src/main/java/org/graylog/security/authservice/rest/AuthServiceBackendsResource.java
@@ -16,6 +16,8 @@
  */
 package org.graylog.security.authservice.rest;
 
+import com.google.common.collect.ImmutableMap;
+import com.google.common.collect.Maps;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
@@ -32,8 +34,14 @@ import org.graylog2.plugin.rest.ValidationFailureException;
 import org.graylog2.plugin.rest.ValidationResult;
 import org.graylog2.rest.PaginationParameters;
 import org.graylog2.rest.models.PaginatedResponse;
+import org.graylog2.search.SearchQuery;
+import org.graylog2.search.SearchQueryField;
+import org.graylog2.search.SearchQueryParser;
 import org.graylog2.shared.rest.resources.RestResource;
 import org.graylog2.shared.security.RestPermissions;
+import org.graylog2.users.PaginatedUserService;
+import org.graylog2.users.RoleService;
+import org.graylog2.users.UserOverviewDTO;
 
 import javax.inject.Inject;
 import javax.validation.constraints.NotBlank;
@@ -42,6 +50,7 @@ import javax.ws.rs.BadRequestException;
 import javax.ws.rs.BeanParam;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
+import javax.ws.rs.DefaultValue;
 import javax.ws.rs.GET;
 import javax.ws.rs.NotFoundException;
 import javax.ws.rs.POST;
@@ -49,9 +58,14 @@ import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
+import javax.ws.rs.QueryParam;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import java.util.Collections;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+import java.util.stream.Collectors;
 
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Strings.isNullOrEmpty;
@@ -62,17 +76,31 @@ import static com.google.common.base.Strings.isNullOrEmpty;
 @Api(value = "System/Authentication/Services/Backends", description = "Manage authentication service backends")
 @RequiresAuthentication
 public class AuthServiceBackendsResource extends RestResource {
+    private static final ImmutableMap<String, SearchQueryField> SEARCH_FIELD_MAPPING = ImmutableMap.<String, SearchQueryField>builder()
+            .put(UserOverviewDTO.FIELD_USERNAME, SearchQueryField.create(UserOverviewDTO.FIELD_USERNAME))
+            .put(UserOverviewDTO.FIELD_FULL_NAME, SearchQueryField.create(UserOverviewDTO.FIELD_FULL_NAME))
+            .put(UserOverviewDTO.FIELD_EMAIL, SearchQueryField.create(UserOverviewDTO.FIELD_EMAIL))
+            .build();
+
     private final DBAuthServiceBackendService dbService;
     private final GlobalAuthServiceConfig globalAuthServiceConfig;
+    private final PaginatedUserService userService;
+    private final RoleService roleService;
     private final AuthServiceBackendUsageCheck usageCheck;
+    private final SearchQueryParser userSearchQueryParser;
 
     @Inject
     public AuthServiceBackendsResource(DBAuthServiceBackendService dbService,
                                        GlobalAuthServiceConfig globalAuthServiceConfig,
+                                       PaginatedUserService userService,
+                                       RoleService roleService,
                                        AuthServiceBackendUsageCheck usageCheck) {
         this.dbService = dbService;
         this.globalAuthServiceConfig = globalAuthServiceConfig;
+        this.userService = userService;
+        this.roleService = roleService;
         this.usageCheck = usageCheck;
+        this.userSearchQueryParser = new SearchQueryParser(UserOverviewDTO.FIELD_FULL_NAME, SEARCH_FIELD_MAPPING);
     }
 
     @GET
@@ -136,6 +164,55 @@ public class AuthServiceBackendsResource extends RestResource {
             throw new BadRequestException("Authentication service backend <" + backendId + "> is still in use");
         }
         dbService.delete(config.id());
+    }
+
+    @GET
+    @Path("{backendId}/users")
+    @ApiOperation("Get paginated users for active authentication service backend")
+    @RequiresPermissions({RestPermissions.AUTH_SERVICE_GLOBAL_CONFIG_READ, RestPermissions.USERS_LIST})
+    public PaginatedResponse<UserOverviewDTO> getUsers(
+            @ApiParam(name = "page") @QueryParam("page") @DefaultValue("1") int page,
+            @ApiParam(name = "per_page") @QueryParam("per_page") @DefaultValue("50") int perPage,
+            @ApiParam(name = "query") @QueryParam("query") @DefaultValue("") String query,
+            @ApiParam(name = "sort", value = "The field to sort the result on", required = true, allowableValues = "username,full_name,email")
+            @DefaultValue(UserOverviewDTO.FIELD_FULL_NAME) @QueryParam("sort") String sort,
+            @ApiParam(name = "order", value = "The sort direction", allowableValues = "asc, desc")
+            @DefaultValue("asc") @QueryParam("order") String order,
+            @ApiParam(name = "backendId", required = true) @PathParam("backendId") @NotBlank String backendId
+    ) {
+        final AuthServiceBackendDTO activeConfig = loadConfig(backendId);
+
+        final PaginatedList<UserOverviewDTO> userList = userService.findPaginatedByAuthServiceBackend(
+                parseSearchQuery(query), page, perPage, sort, order, activeConfig.id());
+
+        return PaginatedResponse.create(
+                "users",
+                userList,
+                query,
+                Collections.singletonMap("roles", createRoleContext(userList.delegate()))
+        );
+    }
+
+    private Map<String, Object> createRoleContext(List<UserOverviewDTO> userList) {
+        final Set<String> roleIds = userList.stream()
+                .flatMap(user -> user.roles().stream())
+                .collect(Collectors.toSet());
+        try {
+            return roleService.findIdMap(roleIds).values()
+                    .stream()
+                    .map(role -> Maps.immutableEntry(role.getId(), Collections.singletonMap("title", role.getName())))
+                    .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
+        } catch (org.graylog2.database.NotFoundException e) {
+            throw new NotFoundException("Couldn't find roles: " + roleIds);
+        }
+    }
+
+    private SearchQuery parseSearchQuery(String query) {
+        try {
+            return userSearchQueryParser.parse(query);
+        } catch (IllegalArgumentException e) {
+            throw new BadRequestException("Invalid argument in search query: " + e.getMessage());
+        }
     }
 
     private boolean checkReadPermission(AuthServiceBackendDTO config) {
